Convert preprocessor parser to functions
